+++ testutil/Fixture/template/custom-template-alt.html.php
@@ -1,0 +1,10 @@
+<?php
+
+use SubstancePHP\HTTP\Renderer\HtmlRenderer;
+
+/**
+ * @var HtmlRenderer $this
+ * @var string $name
+ */
+?>
+<p>OVERRIDE: <?= $this->h($name) ?></p>